Move IPMI messaging to a common file
The existing code to open /dev/ipmi and send IPMI messages, currently
in src/dellssd.c, can be used for sending IPMI messages on AMD
systems. This patch moves the send_ipmi() and ipmi_open() routines
to a common src/ipmi.c file, no functional changes.

This is done in preparation for supporting LED support using IPMI
on some AMD systems.
