+++ src/Makefile.am
@@ -18,11 +18,11 @@
 
 COMMON_SRCS      = ahci.c block.c cntrl.c config_file.c enclosure.c list.c \
                    raid.c scsi.c slave.c status.c sysfs.c smp.c dellssd.c \
-                   utils.c pci_slot.c vmdssd.c udev.c amd_sgpio.c \
+                   utils.c pci_slot.c vmdssd.c udev.c amd_sgpio.c ipmi.c\
                    ahci.h amd_sgpio.h block.h cntrl.h config_file.h dellssd.h \
                    enclosure.h ibpi.h list.h pci_slot.h pidfile.h raid.h scsi.h \
                    ses.h slave.h smp.h status.h sysfs.h udev.h utils.h version.h \
-                   vmdssd.h
+                   vmdssd.h ipmi.h
 LEDMON_SRCS      = ledmon.c pidfile.c $(COMMON_SRCS)
 LEDCTL_SRCS      = ledctl.c $(COMMON_SRCS)
 TEST_CONFIG_SRCS = config_file.c list.c utils.c
